# Reduced the AST traverse calls to improve ECJ performance