@rspilker 2018-07-17
Thanks for the research.

Can you please also add your name to the AUTHORS file in the appropriate location? That way we are legally covered.