@rPraml 2018-07-19
See #1787